# Use minimum tr patterns